backend/requirements.txt
fastapi
uvicorn[standard]
# Herramientas para Data Science y el modelo de riesgo
pandas
scikit-learn
numpy
# Opcionales, datos climáticos de la NASA (NetCDF, GRIB)
xarray
netCDF4
#Geocodificación y APIs
geopy==2.3.0              # Convertir países o ciudades a coordenadas
requests==2.31.0          # Para hacer peticiones HTTP (Meteomatics API)
#Variables de entorno
python-dotenv==1.0.0      # Para cargar .env con credenciales y configuración
